+++ deliverable/ModbusAP.h
@@ -22,7 +22,7 @@
 int connectToServer(char* ip, int port);
 int disconnectFromServer(int socketfd);
 int readHoldingRegisters(int socketfd, uint16_t id, uint16_t startingAddress,
-                         uint16_t quantity, uint16_t* dataToRead);
+                         uint16_t quantity, uint8_t* dataToRead);
 
 int writeMultipleRegisters(int socketfd, uint16_t id, uint16_t startingAddress,
                            uint16_t quantity, uint16_t* data);
